Moved test data for dapp to a new file
